+++ src-console/ConsoleApp_net452_CosmosDb/Entities/EntityReading.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace ConsoleAppCosmosDb.Entities
+{
+    [Serializable]
+    public class EntityReading
+    {
+        public string Status { get; set; }
+        public int Weight { get; set; }
+        public EntityReadingValue Value { get; set; }
+    }
+}