+++ src-console/ConsoleApp_net452_CosmosDb/Entities/EntityReadingValue.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace ConsoleAppCosmosDb.Entities
+{
+    [Serializable]
+    public class EntityReadingValue
+    {
+        public int Weight { get; set; }
+    }
+}