+++ src/System.Linq.Dynamic.Core/DynamicClass.net35.cs
@@ -1,4 +1,6 @@
 ﻿#if NET35
+using System.Collections.Generic;
+
 namespace System.Linq.Dynamic.Core
 {
     /// <summary>