+++ src/System.Linq.Dynamic.Core/ParsingConfig.cs
@@ -22,6 +22,14 @@ namespace System.Linq.Dynamic.Core
             EvaluateGroupByAtDatabase = true
         };
 
+        /// <summary>
+        /// Default ParsingConfig for CosmosDb
+        /// </summary>
+        public static ParsingConfig DefaultCosmosDb { get; } = new ParsingConfig
+        {
+            RenameEmptyParameterExpressionNames = true
+        };
+
         private IDynamicLinkCustomTypeProvider _customTypeProvider;
 
         private IExpressionPromoter _expressionPromoter;
@@ -96,13 +104,6 @@ namespace System.Linq.Dynamic.Core
         public bool AreContextKeywordsEnabled { get; set; } = true;
 
         /// <summary>
-        /// Gets or sets a value indicating whether to use dynamic object class for anonymous types.
-        ///
-        /// Default value is false.
-        /// </summary>
-        public bool UseDynamicObjectClassForAnonymousTypes { get; set; } = false;
-
-        /// <summary>
         /// Gets or sets a value indicating whether the EntityFramework version supports evaluating GroupBy at database level.
         /// See https://docs.microsoft.com/en-us/ef/core/what-is-new/ef-core-2.1#linq-groupby-translation
         /// 
@@ -133,6 +134,13 @@ namespace System.Linq.Dynamic.Core
         /// Default value is false.
         /// </summary>
         public bool RenameParameterExpression { get; set; } = false;
+
+        /// <summary>
+        /// Prevents any System.Linq.Expressions.ParameterExpression.Name value from being empty by substituting a random 16 character word.
+        /// 
+        /// Default value is false.
+        /// </summary>
+        public bool RenameEmptyParameterExpressionNames { get; set; } = false;
 
         /// <summary>
         /// By default when a member is not found in a type and the type has a string based index accessor it will be parsed as an index accessor. Use