+++ test/System.Linq.Dynamic.Core.Tests/Parser/ExpressionParserTests.cs
@@ -1,6 +1,11 @@
-﻿using System.Linq.Dynamic.Core.Parser;
-using System.Linq.Expressions;
+﻿using Moq;
 using NFluent;
+using System.Collections.Generic;
+using System.Linq.Dynamic.Core.CustomTypeProviders;
+using System.Linq.Dynamic.Core.Exceptions;
+using System.Linq.Dynamic.Core.Parser;
+using System.Linq.Dynamic.Core.Tests.Entities;
+using System.Linq.Expressions;
 using Xunit;
 
 namespace System.Linq.Dynamic.Core.Tests.Parser
@@ -65,6 +70,92 @@ namespace System.Linq.Dynamic.Core.Tests.Parser
 
             // Act
             var parsedExpression = sut.Parse(null).ToString();
+
+            // Assert
+            Check.That(parsedExpression).Equals(result);
+        }
+
+        [Theory]
+        [InlineData("string(\"\")", "")]
+        [InlineData("string(\"a\")", "a")]
+        [InlineData("int(42)", 42)]
+        public void Parse_CastStringIntShouldReturnConstantExpression(string expression, object result)
+        {
+            // Arrange
+            ParameterExpression[] parameters = { ParameterExpressionHelper.CreateParameterExpression(typeof(bool), "x") };
+            var sut = new ExpressionParser(parameters, expression, null, null);
+
+            // Act
+            var constantExpression = (ConstantExpression)sut.Parse(null);
+
+            // Assert
+            Check.That(constantExpression.Value).Equals(result);
+        }
+
+        [Theory]
+#if NET452
+        [InlineData("int?(5)", typeof(int?), "Convert(5)")]
+        [InlineData("int?(null)", typeof(int?), "Convert(null)")]
+        [InlineData("string(null)", typeof(string), "Convert(null)")]
+#else
+        [InlineData("int?(5)", typeof(int?), "Convert(5, Nullable`1)")]
+        [InlineData("int?(null)", typeof(int?), "Convert(null, Nullable`1)")]
+        [InlineData("string(null)", typeof(string), "Convert(null, String)")]
+#endif
+        public void Parse_NullableShouldReturnNullable(string expression, object resultType, object result)
+        {
+            // Arrange
+            ParameterExpression[] parameters = { ParameterExpressionHelper.CreateParameterExpression(typeof(bool), "x") };
+            var sut = new ExpressionParser(parameters, expression, null, null);
+
+            // Act
+            var unaryExpression = (UnaryExpression)sut.Parse(null);
+
+            // Assert
+            Check.That(unaryExpression.Type).Equals(resultType);
+            Check.That(unaryExpression.ToString()).Equals(result);
+        }
+
+        private readonly ParsingConfig _parsingConfig;
+        private readonly Mock<IDynamicLinkCustomTypeProvider> _dynamicTypeProviderMock;
+
+        public ExpressionParserTests()
+        {
+            _dynamicTypeProviderMock = new Mock<IDynamicLinkCustomTypeProvider>();
+            _dynamicTypeProviderMock.Setup(dt => dt.GetCustomTypes()).Returns(new HashSet<Type>() { typeof(Company), typeof(MainCompany) });
+            _dynamicTypeProviderMock.Setup(dt => dt.ResolveType(typeof(Company).FullName)).Returns(typeof(Company));
+            _dynamicTypeProviderMock.Setup(dt => dt.ResolveType(typeof(MainCompany).FullName)).Returns(typeof(MainCompany));
+            _dynamicTypeProviderMock.Setup(dt => dt.ResolveTypeBySimpleName("Company")).Returns(typeof(Company));
+            _dynamicTypeProviderMock.Setup(dt => dt.ResolveTypeBySimpleName("MainCompany")).Returns(typeof(MainCompany));
+
+            _parsingConfig = new ParsingConfig
+            {
+                CustomTypeProvider = _dynamicTypeProviderMock.Object
+            };
+        }
+
+        [Theory]
+        [InlineData("it.MainCompany.Name != null", "(company.MainCompany.Name != null)")]
+        [InlineData("@MainCompany.Companies.Count() > 0", "(company.MainCompany.Companies.Count() > 0)")]
+        [InlineData("Company.Equals(null, null)", "Equals(null, null)")]
+        [InlineData("MainCompany.Name", "company.MainCompany.Name")]
+        [InlineData("Company.Name", "No property or field 'Name' exists in type 'Company'")]
+        public void Parse_PrioritizePropertyOrFieldOverTheType(string expression, string result)
+        {
+            // Arrange
+            ParameterExpression[] parameters = { ParameterExpressionHelper.CreateParameterExpression(typeof(Company), "company") };
+            var sut = new ExpressionParser(parameters, expression, null, _parsingConfig);
+
+            // Act
+            string parsedExpression = null;
+            try
+            {
+                parsedExpression = sut.Parse(null).ToString();
+            }
+            catch (ParseException e)
+            {
+                parsedExpression = e.Message;
+            }
 
             // Assert
             Check.That(parsedExpression).Equals(result);